Add files via upload
The first attempt at the coding of a battle system for the game! 

The main aim at this stage was to get functional code not have the game balenced correctly as this will interact with the economic system when it is introduced, simply change the variables at the top to modify the battle. As for now everything shhots at what it is supposed too, I truncated to 50 rounds thus far as it can end up in a near infinte loop in certain situations. Have fun with it, later I will account for multiple turns and border motion!